+++ fs_ext/src/tokio/fsx/dir/checks/size.rs
@@ -3,7 +3,7 @@ use {
     std::{io, path::Path},
 };
 
-pub async fn size(path: impl AsRef<Path>) -> io::Result<u128> {
+pub async fn size(path: impl AsRef<Path>) -> io::Result<u64> {
     let path = path.as_ref().to_owned();
     asyncify(move || fsx::dir::size(path)).await
 }
